Add editable styles of text
