+++ frontend/src/Components/AboutMe/AboutMe.css
@@ -16,6 +16,7 @@
 .card-element {
 	max-width: 40%;
 	margin-right: 2em;
+	margin-top: 2em;
 }
